refact

+++ cloud-network/src/main/java/ru/lexp00/storage/cloud/network/server/ProcessingHandler.java
@@ -2,17 +2,16 @@ package ru.lexp00.storage.cloud.network.server;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
-import org.w3c.dom.ls.LSException;
 import ru.lexp00.storage.cloud.network.common.*;
 
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.*;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.util.stream.Stream;
 
 public class ProcessingHandler extends ChannelInboundHandlerAdapter {
+
+
+    private final String prefixDir = "[DIR]";
+    private final String splitDelimiter = " ";
 
     private final String serverDir = "./cloud-server";
 
@@ -32,7 +31,7 @@ public class ProcessingHandler extends ChannelInboundHandlerAdapter {
 
         } else if (msg instanceof DirMessage) {
             System.out.println("Прилетело сообщение из декодера о создании папки");
-            String dirTitle = ((DirMessage) msg).getDirTitle();
+            String dirTitle = ((DirMessage) msg).getTitleDir();
             Path path = Paths.get(serverDir, serverFiles, dirTitle);
             if (!Files.exists(path)) {
                 Files.createDirectory(path);
@@ -43,27 +42,11 @@ public class ProcessingHandler extends ChannelInboundHandlerAdapter {
             sendMessageWithListFilesOnServer(ctx);
 
         } else if (msg instanceof RenameMessage) {
-            System.out.println("Прилетело сообщение из декодера с задаче переименовать папку на сервере");
             RenameMessage renameMessage = (RenameMessage) msg;
             String lastTitleFile = renameMessage.getLastTitleFile();
-            System.out.println("lastTitleFile: " + lastTitleFile);
-
             String newTitleFile = renameMessage.getNewTitleFile();
-            System.out.println("newTitleFile: " + newTitleFile);
-
-
             Path destinationPath = Paths.get(serverDir, serverFiles, newTitleFile);
-            Path sourcePath;
-            if (!lastTitleFile.contains("[DIR]")) {
-                sourcePath = Paths.get(serverDir, serverFiles, lastTitleFile);
-                System.out.println("SourcePath: " + sourcePath);
-            } else {
-                String[] strLasFile = lastTitleFile.split(" ");
-                sourcePath = Paths.get(serverDir, serverFiles, strLasFile[1]);
-                System.out.println("SourcePath: " + sourcePath);
-            }
-            System.out.println("DestinationPath" + destinationPath);
-            System.out.println(Files.exists(sourcePath));
+            Path sourcePath = getPath(lastTitleFile);
             if (!Files.exists(destinationPath)) {
                 try {
                     Files.move(sourcePath, destinationPath, StandardCopyOption.REPLACE_EXISTING);
@@ -73,62 +56,48 @@ public class ProcessingHandler extends ChannelInboundHandlerAdapter {
             }
             sendMessageWithListFilesOnServer(ctx);
         } else if (msg instanceof DeleteMessage) {
-            System.out.println("Пришло сообщение из декодера с задачей об удалении файла на сервере");
             DeleteMessage deleteMessage = (DeleteMessage) msg;
-            String strFile = deleteMessage.getStrTitle();
+            String titleFile = deleteMessage.getTitleFile();
             Path path;
-            if (!strFile.contains("[DIR]")) {
-                path = Paths.get(serverDir, serverFiles, strFile);
+            path = getPath(titleFile);
+            if (!Files.exists(path)) {
+                Files.delete(path);
             } else {
-                String[] str = strFile.split(" ");
-                path = Paths.get(serverDir, serverFiles, str[1]);
+                throw new RuntimeException("Такой файл не найден");
             }
-            Files.delete(path);
             sendMessageWithListFilesOnServer(ctx);
         } else if (msg instanceof FileMessage) {
             System.out.println("Пришло сообщение с файлом из декодера");
             FileMessage fileMessage = (FileMessage) msg;
-            String strTitle = fileMessage.getFileTitle();
-
+            String strTitle = fileMessage.getTitleFile();
             Path path;
-            if (!strTitle.contains("[DIR]")) {
+            if (!strTitle.contains(prefixDir)) {
                 path = Paths.get(serverDir, serverFiles, strTitle);
-                Files.write(path, fileMessage.getData());
+                Files.write(path, fileMessage.getDataFile());
             } else {
-                //todo добавить копирование папок
-
-//                String[] str = strTitle.split(" ");
-//                Path rootPath = Paths.get(serverDir, serverFiles, str[1]);
-//                try {
-//                    Files.walkFileTree(rootPath, new SimpleFileVisitor<Path>() {
-//                        @Override
-//                        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-//                            System.out.println("delete file: " + file.toString());
-//                            Files.createDirectory(file);
-//                            return FileVisitResult.CONTINUE;
-//                        }
-//
-//                        @Override
-//                        public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-//                            Files.createDirectory(dir);
-//                            System.out.println("delete dir: " + dir.toString());
-//                            return FileVisitResult.CONTINUE;
-//                        }
-//
-//                    });
-//                } catch(IOException e){
-//                    e.printStackTrace();
-//            }
+                throw new RuntimeException("На данный момент не умеем копировать папки");
             }
             sendMessageWithListFilesOnServer(ctx);
         } else if (msg instanceof FileRequest) {
             FileRequest fileRequest = (FileRequest) msg;
-            String strFile = fileRequest.getStrFile();
+            String strFile = fileRequest.getTitleFile();
             Path path = Paths.get(serverDir, serverFiles, strFile);
             FileMessage fileMessage = new FileMessage(path, State.SEND_FILE);
             ctx.writeAndFlush(fileMessage);
-            System.out.println("Server Handler: отправил запрос в Server Encoder");
+        } else {
+            throw new RuntimeException("Не известный формат сообщения " + msg.getClass().getCanonicalName());
         }
+    }
+
+    private Path getPath(String titleFile) {
+        Path path;
+        if (!titleFile.contains(prefixDir)) {
+            path = Paths.get(serverDir, serverFiles, titleFile);
+        } else {
+            String[] str = titleFile.split(splitDelimiter);
+            path = Paths.get(serverDir, serverFiles, str[1]);
+        }
+        return path;
     }
 
     @Override
@@ -145,7 +114,6 @@ public class ProcessingHandler extends ChannelInboundHandlerAdapter {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-        System.out.println("Отправили сообщение со всем списком файлов на сервере");
     }
 }
